RISC-V: Add auxiliary files
To support a new architecture we need to provide a few helper functions
for memory, cache, timer, etc support.

This patch adds the remainders of those. Some bits are still disabled,
as I couldn't guarantee that we're always running on models / in modes
where the respective hardware is available.

Signed-off-by: Alexander Graf <agraf@suse.de>

---

v2 -> v3:

  - Fix riscv32 target

v3 -> v4:

  - Change copyright from 2013 to 2018
  - Fix coding style
  - Resurrect time reading code
  - Add fence.i for icache flush

v4 -> v5:
  - Mark grub_arch_sync_dma_caches args as unused
  - Whitespace fix