travis: Disable IA64 target
The IA64 target currently does not compile due to the following error
message:

  grub-mkimage: error: undefined symbol grub_arch_sync_dma_caches.

If anyone cares enough about IA64 to fix it up, be my guest and revert
this commit afterwards. For now, we really need to move forward with
a fully successful travis run to ensure that we catch regressions early
on.

Signed-off-by: Alexander Graf <agraf@csgraf.de>